Move the database feature checks out of common, behind a databases barrel
