+++ enterprise/frontend/src/metabase-enterprise/database_routing/DatabaseRoutingSection/utils.tsx
@@ -4,10 +4,10 @@ import { t } from "ttag";
 import { Link } from "metabase/common/components/Link";
 import {
   hasActionsEnabled,
-  hasFeature,
   hasTableEditingEnabled,
   hasWritableConnectionDetails,
 } from "metabase/common/utils/database";
+import { hasFeature } from "metabase/databases";
 import { Text } from "metabase/ui";
 import type { Database } from "metabase-types/api";
