+++ frontend/src/metabase/common/utils/database.ts
@@ -1,13 +1,7 @@
 import { t } from "ttag";
 
 import { SAVED_QUESTIONS_VIRTUAL_DB_ID } from "metabase-lib/v1/metadata/utils/saved-questions";
-import type {
-  Card,
-  Dashboard,
-  Database,
-  DatabaseFeature,
-  DatabaseId,
-} from "metabase-types/api";
+import type { Card, Dashboard, Database, DatabaseId } from "metabase-types/api";
 
 export const isDbModifiable = (
   database:
@@ -35,36 +29,6 @@ export const getDbNotModifiableMessage = (
     : // eslint-disable-next-line metabase/no-literal-metabase-strings -- admin-only: must name Metabase Cloud to distinguish it from a whitelabeled instance
       t`This database is managed by Metabase Cloud and cannot be modified.`;
 };
-
-export const hasFeature = (
-  database: Pick<Database, "features">,
-  feature: DatabaseFeature,
-) => {
-  return database.features?.includes(feature) ?? false;
-};
-
-/**
- * The v1 `Database.hasFeature` treated "join" as any join type. It is not a
- * `DatabaseFeature` of its own, so it needs naming here.
- */
-const JOIN_FEATURES: DatabaseFeature[] = [
-  "left-join",
-  "right-join",
-  "inner-join",
-  "full-join",
-];
-
-export const supportsJoins = (database: Pick<Database, "features">) =>
-  JOIN_FEATURES.some((feature) => hasFeature(database, feature));
-
-/**
- * For features that are optional on the thing requiring them, such as an
- * expression clause with no `requiresFeature`. No requirement means supported.
- */
-export const hasRequiredFeature = (
-  database: Pick<Database, "features">,
-  feature: DatabaseFeature | null | undefined,
-) => feature == null || hasFeature(database, feature);
 
 export const hasActionsEnabled = (database: Pick<Database, "settings">) => {
   return Boolean(database.settings?.["database-enable-actions"]);
