+++ frontend/src/metabase/common/utils/database.unit.spec.ts
@@ -6,9 +6,7 @@ import {
   dashboardUsesRoutingEnabledDatabases,
   findDatabaseByName,
   hasDbRoutingEnabled,
-  hasRequiredFeature,
   questionUsesRoutingEnabledDatabase,
-  supportsJoins,
 } from "./database";
 
 describe("database routing utility functions", () => {
@@ -148,36 +146,5 @@ describe("findDatabaseByName", () => {
     expect(
       findDatabaseByName([...databases, virtualDb], "Saved Questions"),
     ).toBeUndefined();
-  });
-});
-
-describe("supportsJoins", () => {
-  it.each(["left-join", "right-join", "inner-join", "full-join"] as const)(
-    "treats %s as join support",
-    (feature) => {
-      expect(supportsJoins(createMockDatabase({ features: [feature] }))).toBe(
-        true,
-      );
-    },
-  );
-
-  it("is false when the database supports no join type", () => {
-    expect(
-      supportsJoins(createMockDatabase({ features: ["expressions"] })),
-    ).toBe(false);
-  });
-});
-
-describe("hasRequiredFeature", () => {
-  const database = createMockDatabase({ features: ["expressions"] });
-
-  it("is true when nothing is required", () => {
-    expect(hasRequiredFeature(database, null)).toBe(true);
-    expect(hasRequiredFeature(database, undefined)).toBe(true);
-  });
-
-  it("defers to the feature when one is required", () => {
-    expect(hasRequiredFeature(database, "expressions")).toBe(true);
-    expect(hasRequiredFeature(database, "nested-queries")).toBe(false);
   });
 });
