+++ frontend/src/metabase/query_builder/components/view/ViewHeader/utils.ts
@@ -1,4 +1,4 @@
-import { hasFeature } from "metabase/common/utils/database";
+import { hasFeature } from "metabase/databases";
 import * as Lib from "metabase-lib";
 import type Question from "metabase-lib/v1/Question";
