+++ frontend/src/metabase/querying/expressions/diagnostics/expression/check-supported-functions.ts
@@ -1,6 +1,6 @@
 import { t } from "ttag";
 
-import { hasRequiredFeature } from "metabase/common/utils/database";
+import { hasRequiredFeature } from "metabase/databases";
 import * as Lib from "metabase-lib";
 import type Metadata from "metabase-lib/v1/metadata/Metadata";
